refactor(storybook): scope storybook specific APIs into a storybook module
